# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 62782285c97c70218d654808
  title: j2ee
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
...
  /user:
    post:
      x-skaffolder-id: 62782286c97c70218d654815
      x-skaffolder-name: create
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                phone:
                  type: number
                username:
                  type: string
    get:
      x-skaffolder-id: 62782286c97c70218d654818
      x-skaffolder-name: list
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 62782286c97c70218d65481c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 62782286c97c70218d654817
      x-skaffolder-name: get
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 62782286c97c70218d654819
      x-skaffolder-name: update
      x-skaffolder-id-resource: 62782285c97c70218d654812
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
...
  /adherent:
    post:
      x-skaffolder-id: 6278270cc97c70218d65482c
      x-skaffolder-name: create
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - adherent
      parameters:
        - name: obj
          x-skaffolder-type: adherent
          in: path
          description: Object to insert
          required: true
          schema:
            type: adherent
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/adherent"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                phone:
                  type: number
    get:
      x-skaffolder-id: 62782726c97c70218d654834
      x-skaffolder-name: list
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF adherent
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - adherent
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/adherent"
  "/adherent/{id}":
    delete:
      x-skaffolder-id: 6278271ac97c70218d654830
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - adherent
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id adherent
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 62782722c97c70218d654832
      x-skaffolder-name: get
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: adherent
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - adherent
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id adherent
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/adherent"
    post:
      x-skaffolder-id: 62782714c97c70218d65482e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: adherent
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - adherent
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id adherent
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/adherent"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/adherent"
  "/adherent/findByname/{key}":
    get:
      x-skaffolder-id: 6278273bc97c70218d654837
      x-skaffolder-name: findByname
      x-skaffolder-id-resource: 627826cac97c70218d65482a
      x-skaffolder-resource: adherent
      x-skaffolder-crudAction: findByname
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByname
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByname/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByname
      tags:
        - adherent
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource name to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/adherent"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      x-skaffolder-id: 62782285c97c70218d654812
      x-skaffolder-id-db: 62782285c97c70218d654809
      x-skaffolder-id-entity: 62782285c97c70218d65480a
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480d
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480e
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480c
          x-skaffolder-type: String
          x-skaffolder-required: true
        phone:
          type: number
          x-skaffolder-id-attr: 62782686c97c70218d654826
          x-skaffolder-type: Number
        username:
          type: string
          x-skaffolder-id-attr: 62782285c97c70218d65480b
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    adherent:
      x-skaffolder-id: 627826cac97c70218d65482a
      x-skaffolder-id-db: 62782285c97c70218d654809
      x-skaffolder-id-entity: 627826cac97c70218d654828
      x-skaffolder-url: /adherent
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 62782741c97c70218d65483a
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 627826cac97c70218d654829
          x-skaffolder-type: String
        phone:
          type: number
          x-skaffolder-id-attr: 62782741c97c70218d654839
          x-skaffolder-type: Number
      x-skaffolder-relations: 
      required:
        - _id
...